refactor: move about into new module, separate AboutSection
